chore: clean up templates for TACC/Core-CMS#710 (#209)
* chore: remove favicon, forget template inheritance

Core-CMS will load the favicon in assets_core.

Core-CMS-Custom is not supposed to overwrite assets_core.

https://github.com/TACC/Core-CMS/pull/710/commits/4086324

* chore: delete unused template

No reason to show the non-loading of a stylesheet.

This template is not loaded, thus the stylesheet is not loaded.

* style: remove extra newline

* chore: remove favicon, forget template inheritance

Core-CMS will load the favicon in assets_core.

Core-CMS-Custom is not supposed to overwrite assets_core.

https://github.com/TACC/Core-CMS/pull/710/commits/4086324

* fix(ecep): needs core-cms v3.12 beta 4